12 Jan 2024, 01:15

+++ squareRootUsing binary search.cpp
@@ -23,7 +23,7 @@ long long int squareRoot(int num,int start,int end,long long int mid)
 
 int main(){
     int num=9;
-    int val=squareRoot(num,1,num-1,-1);
+    long long int val=squareRoot(num,1,num-1,-1);
     cout<<"The value= "<<val<<endl;
     
     return 0;
